+++ src/ej_dataviz/apps.py
@@ -3,3 +3,7 @@ from django.apps import AppConfig
 
 class EjDatavizConfig(AppConfig):
     name = 'ej_dataviz'
+
+    def ready(self):
+        from . import roles
+        self.roles = roles
