+++ src/ej_reports/apps.py
@@ -6,7 +6,10 @@ class EjReportsConfig(AppConfig):
     name = 'ej_reports'
     verbose_name = _('Reports')
     rules = None
+    # roles = None
 
     def ready(self):
         from . import rules
+        # from . import roles
         self.rules = rules
+        # self.roles = roles
